+++ maslov.aleksey/S7/graph.cpp
@@ -1,0 +1,136 @@
+#include "graph.hpp"
+#include <algorithm>
+#include <stdexcept>
+
+void maslov::Graph::extract(const Graph & graph, const std::set< std::string > & extractVertexes)
+{
+  for (auto it = extractVertexes.begin(); it != extractVertexes.end(); it++)
+  {
+    addVertex(*it);
+  }
+  for (auto it = graph.edges.begin(); it != graph.edges.end(); it++)
+  {
+    const auto vertexes = it->first;
+    if (hasVertex(vertexes.first) && hasVertex(vertexes.second))
+    {
+      for (auto itWeight = it->second.begin(); itWeight != it->second.end(); ++itWeight)
+      {
+        addEdge(vertexes.first, vertexes.second, *itWeight);
+      }
+    }
+  }
+}
+
+void maslov::Graph::merge(const Graph & graph)
+{
+  for (auto it = graph.edges.begin(); it != graph.edges.end(); it++)
+  {
+    for (auto itWeight = it->second.begin(); itWeight != it->second.end(); ++itWeight)
+    {
+      addEdge(it->first.first, it->first.second, *itWeight);
+    }
+  }
+  for (auto it = graph.vertexes.begin(); it != graph.vertexes.end(); it++)
+  {
+    addVertex(*it);
+  }
+}
+
+void maslov::Graph::addVertex(const std::string & v)
+{
+  vertexes.insert(v);
+}
+
+bool maslov::Graph::hasVertex(const std::string & vertex) const
+{
+  return vertexes.find(vertex) != vertexes.end();
+}
+
+bool maslov::Graph::hasEdge(const std::string & v1, const std::string & v2, int weight) const
+{
+  auto it = edges.find({v1, v2});
+  if (it == edges.end())
+  {
+    return false;
+  }
+  for (int existWeight: it->second)
+  {
+    if (existWeight == weight)
+    {
+      return true;
+    }
+  }
+  return false;
+}
+
+void maslov::Graph::addEdge(const std::string & v1, const std::string & v2, int weight)
+{
+  edges[{v1, v2}].emplace_back(weight);
+  vertexes.insert(v1);
+  vertexes.insert(v2);
+}
+
+void maslov::Graph::deleteEdge(const std::string & v1, const std::string & v2, int weight)
+{
+  auto it = edges.find({v1, v2});
+  if (it == edges.end())
+  {
+    throw std::logic_error("ERROR: there is no such edge");
+  }
+  auto & weights = it->second;
+  auto itWeight = weights.begin();
+  bool deleted = false;
+  while (itWeight != weights.end())
+  {
+    if (*itWeight == weight)
+    {
+      itWeight = weights.erase(itWeight);
+      deleted = true;
+    }
+    else
+    {
+      ++itWeight;
+    }
+  }
+  if (!deleted)
+  {
+    throw std::logic_error("ERROR: there is no such weight");
+  }
+  if (weights.empty())
+  {
+    edges.erase(it);
+  }
+}
+
+std::vector< std::string > maslov::Graph::getVertexes() const
+{
+  return {vertexes.begin(), vertexes.end()};
+}
+
+std::vector< std::pair< std::string, std::vector< int > > > maslov::Graph::getOutbound(const std::string & vertex) const
+{
+  std::vector< std::pair< std::string, std::vector< int > > > result;
+  for (auto it = edges.begin(); it != edges.end(); it++)
+  {
+    if (it->first.first == vertex)
+    {
+      result.push_back({it->first.second, edges.at({it->first.first, it->first.second})});
+    }
+  }
+  std::sort(result.begin(), result.end());
+  return result;
+}
+
+std::vector< std::pair< std::string, std::vector< int > > > maslov::Graph::getInbound(const std::string & vertex) const
+{
+  std::vector< std::pair< std::string, std::vector< int > > > result;
+  for (auto it = edges.begin(); it != edges.end(); it++)
+  {
+    if (it->first.second == vertex)
+    {
+      result.push_back({it->first.first, edges.at({it->first.first, it->first.second})});
+    }
+  }
+  std::sort(result.begin(), result.end());
+  return result;
+}
