+++ maslov.aleksey/S7/graph.hpp
@@ -5,8 +5,6 @@
 #include <unordered_map>
 #include <set>
 #include <vector>
-#include <algorithm>
-#include <stdexcept>
 
 namespace maslov
 {
@@ -34,141 +32,8 @@ namespace maslov
     void extract(const Graph & graph, const std::set< std::string > & extractVertexes);
    private:
     std::set< std::string > vertexes;
-    std::unordered_map<std::pair< std::string, std::string >, std::vector< int >, PairHash > edges;
+    std::unordered_map< std::pair< std::string, std::string >, std::vector< int >, PairHash > edges;
   };
-
-  void Graph::extract(const Graph & graph, const std::set< std::string > & extractVertexes)
-  {
-    for (auto it = extractVertexes.begin(); it != extractVertexes.end(); it++)
-    {
-      addVertex(*it);
-    }
-    for (auto it = graph.edges.begin(); it != graph.edges.end(); it++)
-    {
-      const auto vertexes = it->first;
-      if (hasVertex(vertexes.first) && hasVertex(vertexes.second))
-      {
-        for (auto itWeight = it->second.begin(); itWeight != it->second.end(); ++itWeight)
-        {
-          addEdge(vertexes.first, vertexes.second, *itWeight);
-        }
-      }
-    }
-  }
-
-  void Graph::merge(const Graph & graph)
-  {
-    for (auto it = graph.edges.begin(); it != graph.edges.end(); it++)
-    {
-      for (auto itWeight = it->second.begin(); itWeight != it->second.end(); ++itWeight)
-      {
-        addEdge(it->first.first, it->first.second, *itWeight);
-      }
-    }
-    for (auto it = graph.vertexes.begin(); it != graph.vertexes.end(); it++)
-    {
-      addVertex(*it);
-    }
-  }
-
-  void Graph::addVertex(const std::string & v)
-  {
-    vertexes.insert(v);
-  }
-
-  bool Graph::hasVertex(const std::string & vertex) const
-  {
-    return vertexes.find(vertex) != vertexes.end();
-  }
-
-  bool Graph::hasEdge(const std::string & v1, const std::string & v2, int weight) const
-  {
-    auto it = edges.find({v1, v2});
-    if (it == edges.end())
-    {
-      return false;
-    }
-    for (int existWeight: it->second)
-    {
-      if (existWeight == weight)
-      {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  void Graph::addEdge(const std::string & v1, const std::string & v2, int weight)
-  {
-    edges[{v1, v2}].emplace_back(weight);
-    vertexes.insert(v1);
-    vertexes.insert(v2);
-  }
-
-  void Graph::deleteEdge(const std::string & v1, const std::string & v2, int weight)
-  {
-    auto it = edges.find({v1, v2});
-    if (it == edges.end())
-    {
-      throw std::logic_error("ERROR: there is no such edge");
-    }
-    auto & weights = it->second;
-    auto itWeight = weights.begin();
-    bool deleted = false;
-    while (itWeight != weights.end())
-    {
-      if (*itWeight == weight)
-      {
-        itWeight = weights.erase(itWeight);
-        deleted = true;
-      }
-      else
-      {
-        ++itWeight;
-      }
-    }
-    if (!deleted)
-    {
-      throw std::logic_error("ERROR: there is no such weight");
-    }
-    if (weights.empty())
-    {
-      edges.erase(it);
-    }
-  }
-
-  std::vector< std::string > Graph::getVertexes() const
-  {
-    return {vertexes.begin(), vertexes.end()};
-  }
-
-  std::vector< std::pair< std::string, std::vector< int > > > Graph::getOutbound(const std::string & vertex) const
-  {
-    std::vector< std::pair< std::string, std::vector< int > > > result;
-    for (auto it = edges.begin(); it != edges.end(); it++)
-    {
-      if (it->first.first == vertex)
-      {
-        result.push_back({it->first.second, edges.at({it->first.first, it->first.second})});
-      }
-    }
-    std::sort(result.begin(), result.end());
-    return result;
-  }
-
-  std::vector< std::pair< std::string, std::vector< int > > > Graph::getInbound(const std::string & vertex) const
-  {
-    std::vector< std::pair< std::string, std::vector< int > > > result;
-    for (auto it = edges.begin(); it != edges.end(); it++)
-    {
-      if (it->first.second == vertex)
-      {
-        result.push_back({it->first.first, edges.at({it->first.first, it->first.second})});
-      }
-    }
-    std::sort(result.begin(), result.end());
-    return result;
-  }
 }
 
 #endif
